jdbc -0- 17/08/2019

+++ src/main/java/hibernate/lesson4/services/HotelService.java
@@ -1,67 +1,51 @@
 package hibernate.lesson4.services;
 
-import hibernate.lesson4.constants.Constants;
 import hibernate.lesson4.dao.HotelDAO;
+import hibernate.lesson4.dao.UserDAO;
+import hibernate.lesson4.exceptions.AuthorizedException;
 import hibernate.lesson4.model.Hotel;
-import org.hibernate.HibernateException;
-import org.hibernate.Session;
-import org.hibernate.SessionFactory;
-import org.hibernate.cfg.Configuration;
-import org.hibernate.query.Query;
 
 import java.util.List;
+
+import static hibernate.lesson4.controller.SessionAuthorization.validate;
 
 public class HotelService {
 
     private HotelDAO hotelDAO = new HotelDAO();
-    private static SessionFactory sessionFactory;
 
     public List<Hotel> findHotelByName(String name) throws Exception {
-        try (Session session = createSessionFactory().openSession()) {
-            Query<Hotel> query = session.createNativeQuery(Constants.FIND_BY_CONTAINED_NAME, Hotel.class);
-            query.setParameter(1, "%" + name + "%");
-
-            return query.list();
-
-        } catch (HibernateException e) {
-            throw new Exception("findHotelByName(String name) method  from class " +
-                    HotelService.class.getName() + " was failed");
-        }
+        validate();
+        return  hotelDAO.findHotelByName(name);
     }
 
     public List<Hotel> findHotelByCity(String city) throws Exception {
-        try (Session session = createSessionFactory().openSession()) {
-            Query<Hotel> query = session.createNativeQuery(Constants.FIND_BY_NAME, Hotel.class);
-            query.setParameter(1, city);
-
-            return query.list();
-
-        } catch (HibernateException e) {
-            throw new Exception("findHotelByCity(String city) method from class " +
-                    HotelService.class.getName() + " was failed");
-        }
+        validate();
+        return hotelDAO.findHotelByCity(city);
     }
 
     public Hotel save(Hotel object) throws Exception {
+        if (!UserDAO.isAdmin())
+            throw new AuthorizedException("user do not have permission in method save(Hotel object) from class " +
+                    HotelService.class.getName());
         return hotelDAO.save(object);
     }
 
     public void delete(long id) throws Exception {
+        if (!UserDAO.isAdmin())
+            throw new AuthorizedException("user do not have permission in method delete(long id) from class " +
+                    HotelService.class.getName());
         hotelDAO.delete(id);
     }
 
     public Hotel update(Hotel object) throws Exception {
+        if (!UserDAO.isAdmin())
+            throw new AuthorizedException("user do not have permission in method update(Hotel object) from class " +
+                    HotelService.class.getName());
         return hotelDAO.update(object);
     }
 
     public Hotel findById(long id) throws Exception {
+        validate();
         return hotelDAO.findById(id);
-    }
-
-    public static SessionFactory createSessionFactory() {
-        if (sessionFactory == null) {
-            sessionFactory = new Configuration().configure().buildSessionFactory();
-        }
-        return sessionFactory;
     }
 }
